// types.ts
export interface Margins {
  top: number;
  bottom: number;
  left: number;
  right: number;
}
// ...
export enum RoofType {
  TUILE_MECANIQUE = "Tuile mécanique",
  TUILE_PLATE = "Tuile plate",
  TUILE_CANAL = "Tuile Canal",
  FIBROCIMENT = "Fibrociment / PST",
}

export enum WindZone {
  ZONE_1 = "Zone 1",
  ZONE_2 = "Zone 2",
  ZONE_3 = "Zone 3",
  ZONE_4 = "Zone 4",
  ZONE_5 = "Zone 5",
}